Modify SubmodelElements to fit with new concept (described in comment of commit)
SM has (multiple) SME(s). One predicate of the SME is the URI (which was the key in the semanticID-tag (see XML)). The object is the value (which was stored in rami:value in the rami:property[] before). The conceptDescription of the property is stored in the subject, the URI identifies.